+++ packages/common/src/plugins/index.ts
@@ -1,3 +1,4 @@
 export * from "./xhrPlugin";
 export * from "./commonErrorPlugin";
 export * from "./fetchPlugin";
+export * from "./unhandledrejectionPlugin";